requirements.txt
attrs==21.2.0
importlib-metadata==4.8.1
isodate==0.6.0
jsonschema==3.2.0
openapi-schema-validator==0.1.5
pyrsistent==0.18.0
PyYAML==5.4.1
six==1.16.0
typing-extensions==3.10.0.2
zipp==3.5.0
compare_objects==1.0.0